pyproject.toml
[project]
name = "deep-research-agent"
version = "0.1.0"
description = "A deep research agent using LangChain and LangGraph"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.0",
    "langgraph>=0.1.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "httpx>=0.27.0",
    "pydantic>=2.8.0",
    "tenacity>=8.3.0",
    "fastapi[standard]>=0.124.4",
    "uvicorn>=0.38.0",
    "langchain-google-genai>=4.0.0",
    "langchain-mcp-adapters>=0.2.1",
    "pymilvus>=2.6.5",
    "langchain-huggingface>=1.2.0",
    "langchain-milvus>=0.3.2",
    "langchain-qdrant>=1.1.0",
    "sentence-transformers>=5.2.0",
    "fastembed>=0.7.4",
    "qdrant-client>=1.16.2",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.0",
    "pymupdf4llm>=0.2.7",
    "pylance>=1.0.0",
    "pytest>=9.0.2",
    "gradio>=6.2.0",
    "pytest-asyncio>=1.3.0",
]
